requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
streamlit-cropper>=0.2.2
